Watchdog
Created a watch dog that will relaunch the server should any of the _exe crash.
This also adds a mini interface that allows the user to launch the server in one button, and similarily close it in one button. It will also only display map_exe which shows all the relevant information