Crab Alpha Mob
Added and implemented crab alpha mob
Please make note of the drop rates on the crab alpha. Where mobs 30+ have a chance to drop a king a pop.
THIS DROP IS NOT EFFECTED BY TH. and is a flat 3% lvls 30-60 and a flat 5% beyond.

With out current TH system, TH would make this a joke, and players would be getting king a pops left and right.